+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add Samsung device detection.
 - Delay parsing `Accept-Language` until `Browser::Base#accept_language` is
   called for the first time.
+- Bump up default size limit for `Accept-Language` and `User-Agent` to 2048
+  bytes.
 
 ## 5.0.0